Creates a resuable component for the permission toggle.
