+++ src/frontend/screens/Onboarding/OnboardingPrivacyPolicy.tsx
@@ -1,17 +1,10 @@
 import * as React from 'react';
-import {
-  View,
-  Text,
-  ScrollView,
-  StyleSheet,
-  TouchableOpacity,
-} from 'react-native';
+import {View, Text, ScrollView, StyleSheet} from 'react-native';
 import {useIntl, defineMessages} from 'react-intl';
 import {PrivacyPolicy} from '../PrivacyPolicy';
-import {BLUE_GREY, WHITE, BLACK, COMAPEO_BLUE} from '../../lib/styles';
-import MaterialIcons from 'react-native-vector-icons/MaterialIcons';
-import {createPersistedPermissionStore} from '../../hooks/persistedState/usePersistedPermission';
-import {useMetrics} from '../../contexts/MetricsContext';
+import {BLUE_GREY, WHITE} from '../../lib/styles';
+import {PermissionToggle} from '../../sharedComponents/PermissionToggle';
+import {usePermissionToggle} from '../../hooks/usePermissionToggle';
 
 const m = defineMessages({
   navTitle: {
@@ -22,42 +15,22 @@ const m = defineMessages({
     id: 'screens.OnboardingPrivacyPolicy.permissionsTitle',
     defaultMessage: 'Current Permissions',
   },
-  shareDiagnostics: {
-    id: 'screens.OnboardingPrivacyPolicy.shareDiagnostics',
-    defaultMessage: 'Share Diagnostic Information',
-  },
 });
 
 export const OnboardingPrivacyPolicy = () => {
   const {formatMessage} = useIntl();
-  const {appMetrics, deviceMetrics} = useMetrics();
-
-  const usePersistedPermission = createPersistedPermissionStore(
-    appMetrics,
-    deviceMetrics,
-  );
-  const {isPermissionEnabled, togglePermission} = usePersistedPermission();
+  const {isPermissionEnabled, togglePermission} = usePermissionToggle();
 
   return (
     <ScrollView contentContainerStyle={styles.scrollContent}>
       <PrivacyPolicy />
       <View style={styles.horizontalLine} />
       <Text style={styles.header}>{formatMessage(m.permissionsTitle)}</Text>
-      <View style={styles.permissionBox}>
-        <Text style={styles.permissionText}>
-          {formatMessage(m.shareDiagnostics)}
-        </Text>
-        <TouchableOpacity
-          onPress={togglePermission}
-          hitSlop={{top: 10, bottom: 10, left: 10, right: 10}}
-          style={[
-            styles.checkBox,
-            isPermissionEnabled && styles.checkBoxChecked,
-          ]}>
-          {isPermissionEnabled && (
-            <MaterialIcons name="check" size={18} color={WHITE} />
-          )}
-        </TouchableOpacity>
+      <View style={styles.permissionToggleContainer}>
+        <PermissionToggle
+          isPermissionEnabled={isPermissionEnabled}
+          togglePermission={togglePermission}
+        />
       </View>
     </ScrollView>
   );
@@ -81,31 +54,13 @@ const styles = StyleSheet.create({
     fontWeight: 'bold',
     marginBottom: 20,
   },
-  permissionBox: {
+  permissionToggleContainer: {
     flexDirection: 'row',
     alignItems: 'center',
+    backgroundColor: WHITE,
     padding: 15,
     borderWidth: 1,
     borderColor: BLUE_GREY,
     borderRadius: 10,
-    backgroundColor: WHITE,
-  },
-  permissionText: {
-    fontSize: 16,
-    color: BLACK,
-    flex: 1,
-  },
-  checkBox: {
-    width: 20,
-    height: 20,
-    borderWidth: 1,
-    borderColor: BLUE_GREY,
-    borderRadius: 2,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  checkBoxChecked: {
-    backgroundColor: COMAPEO_BLUE,
-    borderColor: COMAPEO_BLUE,
   },
 });
